# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pragma-api",
    "description": "0D Master API for vaults",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://{environment}.0d.finance/api/v1",
      "variables": {
        "environment": {
          "default": "devnet",
          "enum": [
            "devnet",
            "production"
          ]
        }
      }
    }
  ],
  "paths": {
    "/users/{address}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "get_user_profile",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "User wallet address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users/{address}/vaults/{vault_id}/kpis": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "get_user_kpis",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "User wallet address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User performance KPIs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserKpi"
                }
              }
            }
          },
          "404": {
            "description": "User KPIs not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/vaults/{vault_id}/apr/series": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_apr_series",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Time period for APR series",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7d"
          }
        ],
        "responses": {
          "200": {
            "description": "Historical APR data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultAprSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/apr/summary": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_apr_summary",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apr_basis",
            "in": "query",
            "description": "APR calculation basis",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "nominal"
          }
        ],
        "responses": {
          "200": {
            "description": "APR summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultAprSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AprBasis": {
        "type": "string",
        "enum": [
          "nominal",
          "inflation_adjusted"
        ]
      },
      "AprPoint": {
        "type": "object",
        "required": [
          "t",
          "apr_pct"
        ],
        "properties": {
          "apr_pct": {
            "type": "number",
            "format": "double"
          },
          "t": {
            "type": "string"
          }
        }
      },
...
      "DisplayCurrency": {
        "type": "string",
        "enum": [
          "usd",
          "usdc"
        ]
      },
...
      "UserKpi": {
        "type": "object",
        "required": [
          "as_of",
          "all_time_pnl_usd",
          "unrealized_pnl_usd",
          "realized_pnl_usd",
          "max_drawdown_pct",
          "sharpe",
          "sortino"
        ],
        "properties": {
          "all_time_pnl_usd": {
            "type": "string"
          },
          "as_of": {
            "type": "string",
            "format": "date-time"
          },
          "max_drawdown_pct": {
            "type": "number",
            "format": "double"
          },
          "realized_pnl_usd": {
            "type": "string"
          },
          "sharpe": {
            "type": "number",
            "format": "double"
          },
          "sortino": {
            "type": "number",
            "format": "double"
          },
          "unrealized_pnl_usd": {
            "type": "string"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "address",
          "chain",
          "display_currency"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency"
          }
        }
      },
...
      "VaultAprSeriesResponse": {
        "type": "object",
        "required": [
          "timeframe",
          "points"
        ],
        "properties": {
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AprPoint"
            }
          },
          "timeframe": {
            "type": "string"
          }
        }
      },
      "VaultAprSummaryResponse": {
        "type": "object",
        "required": [
          "apr_pct",
          "apr_basis"
        ],
        "properties": {
          "apr_basis": {
            "$ref": "#/components/schemas/AprBasis"
          },
          "apr_pct": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  },
...
}